% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions for pgfeatureserv Functions}
\arguments{
\item{...}{Unused parameters.}

\item{response}{A flag indicating whether to return the raw response object.}

\item{base_url}{A character string of the base URL.}

\item{path}{A character string of the path or endpoint.}

\item{user}{A character string of the user. This allows whoever is running
the server your are querying to contact you in case of misuse.}

\item{verbose}{A flag indicating whether to provide verbose GET response.}

\item{collection_id}{A character string of the collection id.}

\item{feature_id}{A positive whole number of the feature id.}

\item{function_id}{A character string of the function id.}

\item{limit}{A positive whole number that limits the number of
features in the response.}

\item{offset}{A positive whole number that starts the response at an offset.}

\item{sortby}{A string that sorts the response items by a property.
Default is ascending but property name can be prepended with
'-' to indicate descending.}

\item{bbox}{A vector of four numbers indicating bounding box in lon/lat
(e.g. \code{c(minLon, minLat, maxLon, maxLat)}) that features in response
must intersect.}

\item{filter}{A named vector or list of the filter(s) to apply,
where the list names correspond to column names and the
list values correspond to the desired
value, e.g. \code{list(gnis_name = "Sangan River")}.}

\item{properties}{A vector of strings of the column names to include.
If NULL (default), all columns are retained.}

\item{transform}{A character vector with the name of the
\href{https://github.com/CrunchyData/pg_featureserv/blob/master/config/pg_featureserv.toml.example#L29}{\emph{valid transform}}
function followed by the parameter values (e.g. c("ST_Simplify", 100)).}

\item{precision}{A positive whole number indicating precision
(i.e., number of decimal places) of GeoJSON coordinates.}

\item{groupby}{A string of the property name (i.e., column name) to group
response features by. This would typically be used with an aggregate
transform function.}

\item{parameters}{A named list of the parameters and values for a
specific function.}
}
\description{
Parameter Descriptions for pgfeatureserv Functions
}
\keyword{internal}
