% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{pgf_collection_crs}
\alias{pgf_collection_crs}
\title{Get collection crs}
\usage{
pgf_collection_crs(
  collection_id,
  base_url,
  path,
  nocache = NULL,
  user = gh_user(),
  verbose = FALSE,
  response = FALSE
)
}
\arguments{
\item{collection_id}{A character string of the collection id.}

\item{base_url}{A character string of the base URL.}

\item{path}{A character string of the path or endpoint.}

\item{nocache}{A character string indicating whether to cache results or not.
Possible values include 'true' or 'false' or NULL.}

\item{user}{A character string of the user. This allows whoever is running
the server your are querying to contact you in case of misuse.}

\item{verbose}{A flag indicating whether to provide verbose GET response.}

\item{response}{A flag indicating whether to return the raw response object.}
}
\value{
A string.
}
\description{
Get crs of a collection.
}
\examples{
\dontrun{
base_url <- "https://features.hillcrestgeo.ca/"
path <- "fwa"
collection_id <- "whse_basemapping.fwa_named_streams"
pgf_collection_crs(,
  collection_id = collection_id, base_url = base_url, path = path
)
}
}
\seealso{
Other collections: 
\code{\link{pgf_collection_bbox}()},
\code{\link{pgf_collection_description}()},
\code{\link{pgf_collection_feature}()},
\code{\link{pgf_collection_features}()},
\code{\link{pgf_collection_geometry_type}()},
\code{\link{pgf_collection_properties}()},
\code{\link{pgf_collections}()}
}
\concept{collections}
