% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{pgf_function}
\alias{pgf_function}
\title{Get function}
\usage{
pgf_function(
  function_id,
  base_url,
  path,
  parameters,
  limit = 10000,
  offset = 0,
  sortby = NULL,
  bbox = NULL,
  properties = NULL,
  precision = NULL,
  transform = NULL,
  user = gh_user(),
  verbose = FALSE,
  response = FALSE
)
}
\arguments{
\item{function_id}{A character string of the function id.}

\item{base_url}{A character string of the base URL.}

\item{path}{A character string of the path or endpoint.}

\item{parameters}{A named list of the parameters and values for a
specific function.}

\item{limit}{A positive whole number that limits the number of
features in the response.}

\item{offset}{A positive whole number that starts the response at an offset.}

\item{sortby}{A string that sorts the response items by a property.
Default is ascending but property name can be prepended with
'-' to indicate descending.}

\item{bbox}{A vector of four numbers indicating bounding box in lon/lat
(e.g. \code{c(minLon, minLat, maxLon, maxLat)}) that features in response
must intersect.}

\item{properties}{A vector of strings of the column names to include.
If NULL (default), all columns are retained.}

\item{precision}{A positive whole number indicating precision
(i.e., number of decimal places) of GeoJSON coordinates.}

\item{transform}{A character vector with the name of the
\href{https://github.com/CrunchyData/pg_featureserv/blob/master/config/pg_featureserv.toml.example#L29}{\emph{valid transform}}
function followed by the parameter values (e.g. c("ST_Simplify", 100)).}

\item{user}{A character string of the user. This allows whoever is running
the server your are querying to contact you in case of misuse.}

\item{verbose}{A flag indicating whether to provide verbose GET response.}

\item{response}{A flag indicating whether to return the raw response object.}
}
\value{
A sf object.
}
\description{
Get output of function.
}
\examples{
\dontrun{
function_id <- "fwa_locatealong"
base_url <- "https://features.hillcrestgeo.ca/"
path <- "fwa"

parameters <- list(
  blue_line_key = 356308001,
  downstream_route_measure = 0
)
pgf_function(function_id,
  base_url = base_url,
  path = path,
  parameters = parameters
)
}

}
\seealso{
Other functions: 
\code{\link{pgf_functions}()}
}
\concept{functions}
