% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{pgf_function_parameters}
\alias{pgf_function_parameters}
\title{Get function parameters}
\usage{
pgf_function_parameters(
  function_id,
  base_url,
  path,
  nocache = NULL,
  user = gh_user(),
  verbose = FALSE,
  response = FALSE
)
}
\arguments{
\item{function_id}{A character string of the function id.}

\item{base_url}{A character string of the base URL.}

\item{path}{A character string of the path or endpoint.}

\item{nocache}{A character string indicating whether to cache results or not.
Possible values include 'true' or 'false' or NULL.}

\item{user}{A character string of the user. This allows whoever is running
the server your are querying to contact you in case of misuse.}

\item{verbose}{A flag indicating whether to provide verbose GET response.}

\item{response}{A flag indicating whether to return the raw response object.}
}
\value{
A tibble.
}
\description{
Get information on function parameters.
}
\examples{
\dontrun{
function_id <- "fwa_locatealong"
base_url <- "https://features.hillcrestgeo.ca/"
path <- "fwa"
pgf_function_parameters(
  function_id = function_id, base_url = base_url, path = path
)
}
}
\seealso{
Other functions: 
\code{\link{pgf_function_description}()},
\code{\link{pgf_function_properties}()},
\code{\link{pgf_function_result}()},
\code{\link{pgf_functions}()}
}
\concept{functions}
