% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgfsr-package.R
\docType{package}
\name{pgfeatureserv-package}
\alias{pgfeatureserv}
\alias{pgfeatureserv-package}
\title{pgfeatureserv: Client for pg_featureserv RESTful web service}
\description{
Client for pg_featureserv, a RESTful geospatial feature server for PostGIS.
}
\author{
\strong{Maintainer}: Seb Dalgarno \email{seb.dalgarno@gmail.com} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID})

}
\keyword{internal}
