% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgfsr-package.R
\docType{package}
\name{pgfsr-package}
\alias{pgfsr}
\alias{pgfsr-package}
\title{pgfsr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Seb Dalgarno \email{dalgarno@unbc.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID})

}
\keyword{internal}
