% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{pgf_collections}
\alias{pgf_collections}
\title{Get collections}
\usage{
pgf_collections(base_url, path, user = gh_user(), verbose = FALSE)
}
\arguments{
\item{base_url}{A character string of the base URL.}

\item{path}{A character string of the path or endpoint.}

\item{user}{A character string of the user. This allows whoever is running
the server your are querying to contact you in case of misuse.}

\item{verbose}{A flag indicating whether to provide verbose GET response.}
}
\value{
A tibble.
}
\description{
Get information about the collections.
}
\examples{
\dontrun{
base_url <- "https://features.hillcrestgeo.ca/"
path <- "fwa"
pgf_collections(base_url = base_url, path = path)
}
}
\seealso{
Other collections: 
\code{\link{pgf_collection_features}()}
}
\concept{collections}
